requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1
django-health-check==3.17.0
django-prometheus==2.2.0
djangorestframework==3.14.0
idna==3.4
prometheus-client==0.16.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
spring-config-client==0.2
sqlparse==0.4.3
urllib3==1.26.15
